+++ src/main/java/org/pageseeder/berlioz/servlet/ErrorHandlerServlet.java
@@ -570,7 +570,7 @@ public final class ErrorHandlerServlet extends HttpServlet {
     if (o == null) return HttpServletResponse.SC_OK;
     else if (o instanceof Integer) {
       Integer code  = (Integer)o;
-      return code.intValue();
+      return code;
     } else {
       LOGGER.error("The 'javax.servlet.error.status_code' must contain an Integer, but was of type: {}", o.getClass().getSimpleName());
       return HttpServletResponse.SC_INTERNAL_SERVER_ERROR;
